Use OpaqueType for backend feature types and remove relative monorepo imports
Replace manual toInternal/is* helpers with OpaqueBackendFeature opaque type.
Move internal types from backend-plugin-api to @internal/backend as the
canonical source, with re-exports back for internal consumers.

Made-with: Cursor
