+++ .changeset/refactor-backend-opaque-types.md
@@ -1,0 +1,7 @@
+---
+'@backstage/backend-plugin-api': patch
+'@backstage/backend-app-api': patch
+'@backstage/backend-test-utils': patch
+---
+
+Migrated internal backend feature types to use `OpaqueType` helpers from `@internal/backend`.
